requirements.txt
Django==2.2.6
psycopg2-binary==2.8.3
drf-yasg==1.17.0
djangorestframework==3.10.3
djangorestframework-xml==1.4.0
pycountry==19.8.18
xmltodict==0.12.0
